+++ packages/core/styles/src/packages/composite/pagination/pagination.style.css.ts
@@ -4,6 +4,7 @@ import { varsThemeBase } from "../../../themes";
 export const container = vanillaStyle({
   display: "flex",
   listStyleType: "none",
+  margin: 0,
   padding: 0,
   gap: varsThemeBase.spacing[1],
 });