+++ packages/core/styles/src/packages/composite/table/table.sprinkle.types.ts
@@ -1,10 +1,25 @@
+import { Conditions } from "../../../types";
 import { tableSprinkle } from "./table.sprinkle.css";
 
-type TablePaddingProperties = keyof typeof tableSprinkle.properties.padding;
+const { properties: propertiesTable } = tableSprinkle;
+
+type TablePaddingProperties = keyof typeof propertiesTable.padding;
+type TableRowBackgroundColorProperties =
+  keyof typeof propertiesTable.backgroundColor;
+
+interface TableConditions<T> extends Conditions<T> {
+  rest?: T;
+  focus?: T;
+  active?: T;
+  hover?: T;
+}
 
 export interface TableSprinkle {
   padding?: TablePaddingProperties;
   width?: string;
   maxWidth?: string;
   minWidth?: string;
+  backgroundColor?:
+    | TableRowBackgroundColorProperties
+    | TableConditions<TableRowBackgroundColorProperties>;
 }